+++ R/plot-map.R
@@ -66,7 +66,8 @@ plot_usmap <- function(regions = c("states", "state", "counties", "county"),
 
   # check for ggplot2
   if (!requireNamespace("ggplot2", quietly = TRUE)) {
-    stop("`ggplot2` must be installed to use `plot_usmap`. Use: install.packages(\"ggplot2\") and try again.")
+    stop("`ggplot2` must be installed to use `plot_usmap`.
+         Use: install.packages(\"ggplot2\") and try again.")
   }
 
   # parse parameters
@@ -78,12 +79,15 @@ plot_usmap <- function(regions = c("states", "state", "counties", "county"),
     geom_args[["colour"]] <- "black"
   }
 
-  if (is.null(geom_args[["fill"]])) {
-    geom_args[["fill"]] <- "white"
-  }
-
   if (is.null(geom_args[["size"]])) {
     geom_args[["size"]] <- 0.4
+  }
+
+  # only use "fill" setting if data is not included
+  if (is.null(geom_args[["fill"]]) & nrow(data) == 0) {
+    geom_args[["fill"]] <- "white"
+  } else if (!is.null(geom_args[["fill"]]) & nrow(data) != 0) {
+    warning("`fill` setting is ignored when `data` is provided. Use `fill` to color regions with solid color when no data is being displayed.")
   }
 
   # create polygon layer